Trying to use a type file for timezone
